fix: task executes before its completed
